fix: move hook
